adicionando clase 25

+++ src/main/java/Proyecto_Backend/Proyecto/service/impl/OdontologoService.java
@@ -1,6 +1,8 @@
 package Proyecto_Backend.Proyecto.service.impl;
 
 import Proyecto_Backend.Proyecto.entity.Odontologo;
+import Proyecto_Backend.Proyecto.entity.Paciente;
+import Proyecto_Backend.Proyecto.exception.ResourceNotFoundException;
 import Proyecto_Backend.Proyecto.repository.IOdontologoRepository;
 import Proyecto_Backend.Proyecto.service.IOdontologoService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,7 +25,12 @@ public class OdontologoService implements IOdontologoService {
     @Override
     public Optional<Odontologo> buscarPorId(Integer id) {
 
-        return odontologoRepository.findById(id);
+        Optional<Odontologo> odontologoEncontrado = odontologoRepository.findById(id);
+        if(odontologoEncontrado.isPresent()){
+            return odontologoEncontrado;
+        } else {
+            throw new ResourceNotFoundException("Odontologo no encontrado");
+        }
     }
 
     @Override
@@ -34,7 +41,13 @@ public class OdontologoService implements IOdontologoService {
 
     @Override
     public void eliminarOdontologo(Integer id) {
-        odontologoRepository.deleteById(id);
+
+        Optional<Odontologo> odontologoEncontrado = buscarPorId(id);
+        if(odontologoEncontrado.isPresent()){
+            odontologoRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Odontologo no encontrado");
+        }
     }
 
     @Override
